@@ -1,13 +1,12 @@
 [package]
 name = "counter_icu_datagen"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
-leptos = { version = "0.7.0", features = ["csr"] }
-leptos_meta = { version = "0.7.0" }
+leptos = { version = "0.8.15", features = ["csr"] }
+leptos_meta = { version = "0.8.5" }
 leptos_i18n = { path = "../../../leptos_i18n", default-features = false, features = [
-  "json_files",
   "csr",
   "plurals",
   "format_nums",
@@ -21,22 +20,20 @@ console_error_panic_hook = { version = "0.1" }
 wasm-bindgen = { version = "0.2" }
 
 # "default-features = false" to turn off compiled_data
-icu = { version = "1.5", default-features = false, features = [
+icu = { version = "2.1.1", default-features = false, features = [
   # feature only needed for experimental features (e.g "format_currency")
   "experimental",
 ] }
-icu_provider = "1.5" # for databake
-zerovec = "0.10" # for databake
+icu_provider_baked = "2.1.1" # for databake
+icu_provider = "2.1.1" # for databake
+zerovec = "0.11" # for databake
 
 # only needed for experimental features (e.g "format_currency")
-icu_pattern = "0.2.0" # for databake
+icu_pattern = "0.4.1" # for databake
 
-[package.metadata.leptos-i18n]
-default = "en"
-locales = ["en", "fr"]
-
-[build-dependencies]
-leptos_i18n_build = { path = "../../../leptos_i18n_build" }
+[build-dependencies.leptos_i18n_build]
+path = "../../../leptos_i18n_build"
+features = ["pretty_print", "csr"]
 
 [profile.release]
 opt-level = "z"
