@@ -1,21 +1,21 @@
 [package]
 name = "routing_csr"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-leptos = { version = "0.7.0", features = ["csr"] }
-leptos_meta = { version = "0.7.0" }
+leptos = { version = "0.8.15", features = ["csr"] }
+leptos_meta = { version = "0.8.5" }
 leptos_i18n = { path = "../../../leptos_i18n", features = ["csr"] }
 leptos_i18n_router = { path = "../../../leptos_i18n_router" }
 serde = { version = "1", features = ["derive"] }
 console_error_panic_hook = { version = "0.1" }
 wasm-bindgen = { version = "0.2" }
 
-leptos_router = { version = "0.7.0" }
+leptos_router = { version = "0.8.11" }
 
-[package.metadata.leptos-i18n]
-default = "en"
-locales = ["en", "fr"]
+[build-dependencies.leptos_i18n_build]
+path = "../../../leptos_i18n_build"
+features = ["pretty_print", "csr"]
