@@ -1,7 +1,7 @@
 [package]
 name = "axum_island"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
@@ -9,12 +9,11 @@ crate-type = ["cdylib", "rlib"]
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-axum = { version = "0.7", optional = true }
-leptos = { version = "0.7.3", features = ["islands"] }
-leptos_meta = { version = "0.7.3" }
-leptos_axum = { version = "0.7.3", optional = true }
+axum = { version = "0.8.7", optional = true }
+leptos = { version = "0.8.15", features = ["islands"] }
+leptos_meta = { version = "0.8.5" }
+leptos_axum = { version = "0.8.7", optional = true }
 leptos_i18n = { path = "../../../leptos_i18n", features = [
-  "track_locale_files",
   "islands",
   "dynamic_load",
 ] }
@@ -22,10 +21,10 @@ serde = { version = "1", features = ["derive"] }
 console_error_panic_hook = { version = "0.1", optional = true }
 wasm-bindgen = "0.2"
 simple_logger = "4"
-tokio = { version = "1.35", features = ["rt-multi-thread"], optional = true }
+tokio = { version = "1.48", features = ["rt-multi-thread"], optional = true }
 log = "0.4"
-tower = { version = "0.5.1", optional = true }
-tower-http = { version = "0.6.1", features = ["fs"], optional = true }
+tower = { version = "0.5.2", optional = true }
+tower-http = { version = "0.6.8", features = ["fs"], optional = true }
 
 [features]
 default = ["hydrate", "ssr"]
@@ -33,6 +32,7 @@ hydrate = [
   "dep:console_error_panic_hook",
   "leptos/hydrate",
   "leptos_i18n/hydrate",
+  "leptos_i18n_build/hydrate",
 ]
 ssr = [
   "dep:axum",
@@ -43,11 +43,12 @@ ssr = [
   "leptos/ssr",
   "leptos_meta/ssr",
   "leptos_i18n/axum",
+  "leptos_i18n_build/ssr",
 ]
 
-[package.metadata.leptos-i18n]
-default = "en"
-locales = ["en", "fr"]
+[build-dependencies.leptos_i18n_build]
+path = "../../../leptos_i18n_build"
+features = ["pretty_print", "islands", "dynamic_load"]
 
 [package.metadata.leptos]
 # The name used by wasm-bindgen/cargo-leptos for the JS/WASM bundle. Defaults to the crate name   
