@@ -3,24 +3,41 @@
 #![deny(warnings)]
 //! This crate provide `build.rs` utilities for the `leptos_i18n` crate.
 
-use std::collections::HashSet;
-use std::fmt::{Display, Write};
-use std::fs::{create_dir_all, File};
-use std::io::BufWriter;
-use std::path::PathBuf;
-use std::rc::Rc;
+pub use datamarker::FormatterOptions;
+pub use leptos_i18n_parser::parse_locales::options::{Config, FileFormat, ParseOptions, parser};
 
-pub use datakey::Options;
-use icu_datagen::baked_exporter::BakedExporter;
-use icu_datagen::prelude::DataKey;
-use icu_datagen::{DatagenDriver, DatagenProvider};
-use icu_locid::LanguageIdentifier;
-use icu_provider::DataError;
-use leptos_i18n_parser::parse_locales;
-use leptos_i18n_parser::parse_locales::error::Result;
-use leptos_i18n_parser::parse_locales::locale::{BuildersKeys, Locale};
+use icu_locale::LocaleFallbacker;
+use icu_provider::{DataError, DataMarkerInfo};
+use icu_provider_export::{
+    DataLocaleFamily, DeduplicationStrategy, ExportDriver, ExportMetadata,
+    baked_exporter::{self, BakedExporter},
+};
+use icu_provider_source::SourceDataProvider;
+use leptos_i18n_parser::parse_locales::{
+    ParsedLocales,
+    error::Result,
+    locale::{BuildersKeys, Locale},
+    parse_locales,
+};
+use std::{
+    collections::HashSet,
+    fmt::{Display, Write},
+    fs::{File, create_dir_all},
+    io::BufWriter,
+    path::PathBuf,
+    rc::Rc,
+};
 
-mod datakey;
+mod datamarker;
+pub mod options;
+
+use crate::options::CodegenOptions;
+
+/// Module for custom formatters.
+pub mod formatter {
+    pub use leptos_i18n_parser::formatters::{Formatter, FormatterToTokens};
+    pub use leptos_i18n_parser::utils::Key;
+}
 
 #[derive(Clone)]
 enum EitherIter<A, B> {
@@ -41,37 +58,39 @@ impl<T, A: Iterator<Item = T>, B: Iterator<Item = T>> Iterator for EitherIter<A,
 
 /// Contains informations about the translations.
 pub struct TranslationsInfos {
-    locales: BuildersKeys,
-    paths: Vec<String>,
+    parsed_locales: ParsedLocales,
 }
 
 impl TranslationsInfos {
-    fn parse_inner(dir_path: Option<PathBuf>) -> Result<Self> {
+    fn parse_inner(dir_path: Option<PathBuf>, cfg: Config) -> Result<Self> {
         // We don't really care for warnings, they will already be displayed by the macro
-        let (locales, _, paths) = parse_locales::parse_locales(true, dir_path)?;
+        let parsed_locales = parse_locales(dir_path, cfg)?;
 
-        Ok(TranslationsInfos { locales, paths })
+        Ok(TranslationsInfos { parsed_locales })
     }
 
     /// Parse the translations and obtain informations about them.
-    pub fn parse() -> Result<Self> {
-        Self::parse_inner(None)
+    pub fn parse(cfg: Config) -> Result<Self> {
+        let this = Self::parse_inner(None, cfg)?;
+        Ok(this)
     }
 
     /// Parse the translations at the given directory and obtain informations about them.
-    pub fn parse_at_dir<P: Into<PathBuf>>(dir_path: P) -> Result<Self> {
-        Self::parse_inner(Some(dir_path.into()))
+    pub fn parse_at_dir<P: Into<PathBuf>>(dir_path: P, cfg: Config) -> Result<Self> {
+        Self::parse_inner(Some(dir_path.into()), cfg)
     }
 
     /// Paths to all files containing translations.
-    pub fn files_paths(&self) -> &[String] {
-        &self.paths
+    pub fn files_paths(&self) -> Option<&[String]> {
+        self.parsed_locales.tracked_files.as_deref()
     }
 
-    /// Output "cargo:rerun-if-changed" for all locales files.
+    /// Output "cargo::rerun-if-changed" for all locales files.
     pub fn rerun_if_locales_changed(&self) {
-        for path in &self.paths {
-            println!("cargo:rerun-if-changed={}", path);
+        if let Some(paths) = self.files_paths() {
+            for path in paths {
+                println!("cargo::rerun-if-changed={path}");
+            }
         }
     }
 
@@ -80,7 +99,7 @@ impl TranslationsInfos {
         fn map_locales(locales: &[Locale]) -> impl Iterator<Item = Rc<str>> + '_ {
             locales.iter().map(|locale| locale.name.name.clone())
         }
-        match &self.locales {
+        match &self.parsed_locales.builder_keys {
             BuildersKeys::NameSpaces { namespaces, .. } => {
                 let iter = namespaces
                     .iter()
@@ -106,7 +125,7 @@ impl TranslationsInfos {
                 strings: &locale.strings,
             })
         }
-        match &self.locales {
+        match &self.parsed_locales.builder_keys {
             BuildersKeys::NameSpaces { namespaces, .. } => {
                 let iter = namespaces.iter().map(|ns| {
                     let locales = map_locales(&ns.locales);
@@ -123,7 +142,7 @@ impl TranslationsInfos {
 
     /// Return an iterator containing the name of each namespaces, if any.
     pub fn get_namespaces(&self) -> Option<impl Iterator<Item = Rc<str>> + '_> {
-        match &self.locales {
+        match &self.parsed_locales.builder_keys {
             BuildersKeys::NameSpaces { namespaces, .. } => {
                 let namespaces = namespaces.iter().map(|ns| ns.key.name.clone());
                 Some(namespaces)
@@ -133,89 +152,159 @@ impl TranslationsInfos {
     }
 
     /// Return an iterator containing each locales in the form of `LanguageIdentifier`.
-    pub fn get_locales_langids(&self) -> impl Iterator<Item = LanguageIdentifier> + '_ {
+    pub fn get_locales_langids(&self) -> impl Iterator<Item = DataLocaleFamily> + '_ {
         self.get_locales()
-            .map(|locale| locale.parse::<LanguageIdentifier>().unwrap())
+            .map(|locale| locale.parse::<DataLocaleFamily>().unwrap())
     }
 
-    fn get_icu_keys_inner(&self, used_icu_keys: &mut HashSet<Options>) {
-        match &self.locales {
+    fn get_icu_keys_inner(&self, used_icu_keys: &mut HashSet<FormatterOptions>) {
+        match &self.parsed_locales.builder_keys {
             BuildersKeys::NameSpaces { keys, .. } => {
                 for builder_keys in keys.values() {
-                    datakey::find_used_datakey(builder_keys, used_icu_keys);
+                    datamarker::find_used_datamarker(builder_keys, used_icu_keys);
                 }
             }
             BuildersKeys::Locales { keys, .. } => {
-                datakey::find_used_datakey(keys, used_icu_keys);
+                datamarker::find_used_datamarker(keys, used_icu_keys);
             }
         }
     }
 
-    /// Return the ICU `DataKey` needed by the translations.
-    pub fn get_icu_keys(&self) -> impl Iterator<Item = DataKey> {
+    /// Return the ICU `DataMarker` needed by the translations.
+    pub fn get_icu_keys(&self) -> impl Iterator<Item = DataMarkerInfo> {
         let mut used_icu_keys = HashSet::new();
         self.get_icu_keys_inner(&mut used_icu_keys);
-        datakey::get_keys(used_icu_keys)
+        datamarker::get_markers(used_icu_keys)
     }
 
-    /// Same as `build_datagen_driver` but can be supplied with additional ICU `DataKey`.
+    /// Same as `build_datagen_driver` but can be supplied with additional ICU `DataMarker`.
     pub fn build_datagen_driver_with_data_keys(
         &self,
-        keys: impl IntoIterator<Item = DataKey>,
-    ) -> DatagenDriver {
-        let mut icu_keys: HashSet<DataKey> = self.get_icu_keys().collect();
+        keys: impl IntoIterator<Item = DataMarkerInfo>,
+    ) -> ExportDriver {
+        let mut icu_keys: HashSet<DataMarkerInfo> = self.get_icu_keys().collect();
         icu_keys.extend(keys);
 
         let locales = self.get_locales_langids();
-        DatagenDriver::new()
-            .with_keys(icu_keys)
-            .with_locales_no_fallback(locales, Default::default())
+
+        ExportDriver::new(
+            locales,
+            DeduplicationStrategy::None.into(),
+            LocaleFallbacker::new_without_data(),
+        )
+        .with_markers(icu_keys)
     }
 
     /// Same as `build_datagen_driver` but can be supplied with additional options.
     /// This usefull if you use `t*_format!` and use formatters not used in the translations.
     pub fn build_datagen_driver_with_options(
         &self,
-        keys: impl IntoIterator<Item = Options>,
-    ) -> DatagenDriver {
-        self.build_datagen_driver_with_data_keys(datakey::get_keys(keys))
+        keys: impl IntoIterator<Item = FormatterOptions>,
+    ) -> ExportDriver {
+        self.build_datagen_driver_with_data_keys(datamarker::get_markers(keys))
     }
 
-    /// Build a `DatagenDriver` using the locales and keys needed for the translations.
-    pub fn build_datagen_driver(&self) -> DatagenDriver {
+    /// Build a `ExportDriver` using the locales and keys needed for the translations.
+    pub fn build_datagen_driver(&self) -> ExportDriver {
         self.build_datagen_driver_with_options(std::iter::empty())
     }
 
-    /// Same as `generate_data` but can be supplied additionnal ICU `DataKey`.
-    pub fn generate_data_with_data_keys(
+    /// Same as `generate_data` but can be supplied additionnal ICU `DataMarker`.
+    pub fn generate_data_with_data_markers(
         &self,
         mod_directory: PathBuf,
-        keys: impl IntoIterator<Item = DataKey>,
-    ) -> Result<(), DataError> {
+        keys: impl IntoIterator<Item = DataMarkerInfo>,
+    ) -> Result<ExportMetadata, DataError> {
         // This is'nt really needed, but ICU4X wants the directory to be empty
         // and Rust Analyzer can trigger the build.rs without cleaning the out directory.
         if mod_directory.exists() {
             std::fs::remove_dir_all(&mod_directory).unwrap();
         }
 
-        let exporter = BakedExporter::new(mod_directory, Default::default()).unwrap();
+        let exporter = BakedExporter::new(mod_directory, {
+            let mut options = baked_exporter::Options::default();
+            options.overwrite = true;
+            options.use_internal_fallback = false;
+            options
+        })
+        .unwrap();
 
         self.build_datagen_driver_with_data_keys(keys)
-            .export(&DatagenProvider::new_latest_tested(), exporter)
+            .export(&SourceDataProvider::new(), exporter)
     }
 
     /// Same as `generate_data` but can be supplied additionnal options.
     pub fn generate_data_with_options(
         &self,
         mod_directory: PathBuf,
-        keys: impl IntoIterator<Item = Options>,
-    ) -> Result<(), DataError> {
-        self.generate_data_with_data_keys(mod_directory, datakey::get_keys(keys))
+        keys: impl IntoIterator<Item = FormatterOptions>,
+    ) -> Result<ExportMetadata, DataError> {
+        self.generate_data_with_data_markers(mod_directory, datamarker::get_markers(keys))
     }
 
     /// Generate an ICU datagen at the given mod_directory using the infos from the translations.
-    pub fn generate_data(&self, mod_directory: PathBuf) -> Result<(), DataError> {
+    pub fn generate_data(&self, mod_directory: PathBuf) -> Result<ExportMetadata, DataError> {
         self.generate_data_with_options(mod_directory, std::iter::empty())
+    }
+
+    /// Generate the `i18n` module at the given mod directory
+    pub fn generate_i18n_module(&self, mod_directory: PathBuf) -> Result<()> {
+        self.generate_i18n_module_with_options(mod_directory, CodegenOptions::default())
+    }
+
+    /// Generate the `i18n` module at the given mod directory with options
+    pub fn generate_i18n_module_with_options(
+        &self,
+        mut mod_directory: PathBuf,
+        options: CodegenOptions,
+    ) -> Result<()> {
+        let ts = leptos_i18n_codegen::gen_code(
+            &self.parsed_locales,
+            options.crate_path.as_ref(),
+            false,
+            options.top_level_attributes.as_ref(),
+        )?;
+
+        #[cfg(feature = "pretty_print")]
+        let ts = {
+            let as_file = syn::parse_quote!(#ts);
+            prettyplease::unparse(&as_file)
+        };
+
+        create_dir_all(&mod_directory)?;
+
+        mod_directory.push(options.module_file_name);
+
+        let mut file = File::create(&mod_directory)?;
+
+        use std::io::Write;
+        write!(&mut file, "{ts}")?;
+
+        Ok(())
+    }
+
+    /// Emit the warnings generated when parsing the translations
+    pub fn emit_warnings(&self) {
+        let warnings = self.parsed_locales.diag.warnings();
+
+        for warning in warnings.iter() {
+            println!("cargo::warning={warning}");
+        }
+    }
+
+    /// emit the errors generated when parsing the translations
+    pub fn emit_errors(&self) {
+        let errors = self.parsed_locales.diag.errors();
+
+        for error in errors.iter() {
+            println!("cargo::error={error}");
+        }
+    }
+
+    /// Emit the diagnostics generated when parsing the translations
+    pub fn emit_diagnostics(&self) {
+        self.emit_warnings();
+        self.emit_errors();
     }
 }
 
@@ -278,11 +367,11 @@ fn write_locales_to_dir<'a>(
 }
 
 impl<
-        'a,
-        N: Iterator<Item = NamespaceTranslations<'a, NL>>,
-        NL: Iterator<Item = LocaleTranslations<'a>>,
-        L: Iterator<Item = LocaleTranslations<'a>>,
-    > TranslationsType<N, L>
+    'a,
+    N: Iterator<Item = NamespaceTranslations<'a, NL>>,
+    NL: Iterator<Item = LocaleTranslations<'a>>,
+    L: Iterator<Item = LocaleTranslations<'a>>,
+> TranslationsType<N, L>
 {
     /// Write the translations in the given directory
     pub fn write_to_dir<P: Into<PathBuf>>(self, path: P) -> std::io::Result<()> {
@@ -349,10 +438,10 @@ impl Display for TranslationsFormatter<'_> {
         f.write_char('[')?;
         let mut iter = self.strings.iter();
         if let Some(first) = iter.next() {
-            write!(f, "{:?}", first)?;
+            write!(f, "{first:?}")?;
         }
         for s in iter {
-            write!(f, ",{:?}", s)?;
+            write!(f, ",{s:?}")?;
         }
         f.write_char(']')
     }
