@@ -1,20 +1,36 @@
-use icu_locid::Error as LocidError;
-use icu_plurals::Error as PluralsError;
-use std::{collections::BTreeSet, fmt::Display, num::TryFromIntError, path::PathBuf, rc::Rc};
+use core::panic;
+use icu_locale::ParseError as LocidError;
+use icu_provider::DataError as IcuDataError;
+use proc_macro2::TokenStream;
+use quote::{ToTokens, format_ident, quote};
+use std::{
+    cell::{Cell, Ref, RefCell},
+    collections::BTreeSet,
+    fmt::{Debug, Display},
+    io,
+    num::TryFromIntError,
+    path::PathBuf,
+    rc::Rc,
+};
 
 use super::{locale::SerdeError, ranges::RangeType};
 use crate::{
     parse_locales::cfg_file,
-    utils::key::{Key, KeyPath},
+    utils::{
+        Location,
+        key::{Key, KeyPath},
+    },
 };
 
 #[derive(Debug)]
+#[non_exhaustive]
 pub enum Error {
+    IoError(io::Error),
     InvalidLocale {
         locale: Rc<str>,
         err: LocidError,
     },
-    PluralRulesError(PluralsError),
+    PluralRulesError(IcuDataError),
     CargoDirEnvNotPresent(std::env::VarError),
     ManifestNotFound(std::io::Error),
     ConfigNotPresent,
@@ -58,179 +74,589 @@ pub enum Error {
     },
     ExplicitDefaultInDefault(KeyPath),
     RecursiveForeignKey {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     MissingForeignKey {
         foreign_key: KeyPath,
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     InvalidForeignKey {
         foreign_key: KeyPath,
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     UnknownFormatter {
         name: String,
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     ConflictingPluralRuleType {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     InvalidForeignKeyArgs {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
         err: serde_json::Error,
     },
     InvalidCountArg {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
         foreign_key: KeyPath,
     },
     InvalidCountArgType {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
         foreign_key: KeyPath,
         input_type: RangeType,
         range_type: RangeType,
     },
     CountArgOutsideRange {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
         foreign_key: KeyPath,
         err: TryFromIntError,
     },
     UnexpectedToken {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
         message: String,
     },
     RangeAndPluralsMix {
         key_path: KeyPath,
     },
     PluralsAtNormalKey {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     DisabledFormatter {
-        locale: Key,
-        key_path: KeyPath,
-        formatter: crate::utils::formatter::Formatter,
+        loc: Location,
+        formatter_err: &'static str,
     },
     DisabledPlurals {
-        locale: Key,
-        key_path: KeyPath,
+        loc: Location,
     },
     NoFileFormats,
     MultipleFilesFormats,
     MissingTranslationsURI,
+    InvalidFormatterArgName {
+        loc: Location,
+        name: String,
+        err: String,
+    },
+    InvalidFormatterArg {
+        loc: Location,
+        arg_name: String,
+        arg: Option<String>,
+        err: String,
+    },
+    InvalidFormatter {
+        loc: Location,
+        err: String,
+    },
+    InvalidAttributeName {
+        loc: Location,
+        value: String,
+    },
+    InvalidAttribute {
+        loc: Location,
+        attr_name: String,
+        attr_value: String,
+        err: String,
+    },
+    InvalidForeignKeyArgForAttribute {
+        loc: Location,
+        arg_name: Key,
+        foreign_key: KeyPath,
+    },
+    UnknownLocaleInInherit {
+        loc: &'static panic::Location<'static>,
+        locale: String,
+    },
+    DefaultLocaleCantInherit {
+        loc: &'static panic::Location<'static>,
+    },
+
+    Custom(String),
 }
 
 impl Display for Error {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
+            Error::IoError(err) => <io::Error as Display>::fmt(err, f),
             Error::CargoDirEnvNotPresent(err) => {
-                write!(f, "Error, can't access env variable \"CARGO_MANIFEST_DIR\": {}", err)
+                write!(
+                    f,
+                    "Error, can't access env variable \"CARGO_MANIFEST_DIR\": {err}"
+                )
             }
             Error::ManifestNotFound(err) => {
-                write!(f, "Error accessing cargo manifest (Cargo.toml) : {}", err)
-            },
+                write!(f, "Error accessing cargo manifest (Cargo.toml) : {err}")
+            }
             Error::ConfigNotPresent => {
-                write!(f, "Could not found \"[package.metadata.leptos-i18n]\" in cargo manifest (Cargo.toml)")
+                write!(
+                    f,
+                    "Could not found \"[package.metadata.leptos-i18n]\" in cargo manifest (Cargo.toml)"
+                )
             }
             Error::ConfigFileDeser(err) => {
-                write!(f, "Parsing of cargo manifest (Cargo.toml) failed: {}", err)
+                write!(f, "Parsing of cargo manifest (Cargo.toml) failed: {err}")
             }
             Error::LocaleFileNotFound(errs) => {
                 for (path, err) in errs {
-                    writeln!(f,
-                        "Could not found file {:?} : {}",
-                        path, err
-                    )?;
+                    writeln!(f, "Could not found file {path:?} : {err}")?;
                 }
                 Ok(())
             }
-            Error::LocaleFileDeser { path, err} => write!(f,
-                "Parsing of file {:?} failed: {}",
-                path, err
-            ),
-            Error::RangeParse {
-                range,
-                range_type
-            } => write!(f,
-                "error parsing {:?} as {}", 
-                range, range_type
-            ),
-            Error::DuplicateLocalesInConfig(duplicates) => write!(f,
-                "Found duplicates locales in configuration (Cargo.toml): {:?}", 
-                duplicates
+            Error::LocaleFileDeser { path, err } => {
+                write!(f, "Parsing of file {path:?} failed: {err}")
+            }
+            Error::RangeParse { range, range_type } => {
+                write!(f, "error parsing {range:?} as {range_type}")
+            }
+            Error::DuplicateLocalesInConfig(duplicates) => write!(
+                f,
+                "Found duplicates locales in configuration (Cargo.toml): {duplicates:?}"
             ),
             Error::InvalidBoundEnd {
                 range,
-                range_type: range_type @ (RangeType::F32 | RangeType::F64)
-            } => write!(f,
-                "the range {:?} end bound is invalid, you can't use exclusif range with {}", 
-                range, range_type
+                range_type: range_type @ (RangeType::F32 | RangeType::F64),
+            } => write!(
+                f,
+                "the range {range:?} end bound is invalid, you can't use exclusif range with {range_type}"
             ),
-            Error::InvalidBoundEnd {
-                range,
-                range_type
-            } => write!(f,
-                "the range {:?} end bound is invalid, you can't end before {}::MIN", 
-                range, range_type
+            Error::InvalidBoundEnd { range, range_type } => write!(
+                f,
+                "the range {range:?} end bound is invalid, you can't end before {range_type}::MIN"
             ),
-            Error::ImpossibleRange(range) => write!(f, "the range {:?} is impossible, it end before it starts",
-                range
+            Error::ImpossibleRange(range) => write!(
+                f,
+                "the range {range:?} is impossible, it end before it starts"
             ),
-            Error::DuplicateNamespacesInConfig(duplicates) => write!(f,
-                "Found duplicates namespaces in configuration (Cargo.toml): {:?}", 
-                duplicates
+            Error::DuplicateNamespacesInConfig(duplicates) => write!(
+                f,
+                "Found duplicates namespaces in configuration (Cargo.toml): {duplicates:?}"
             ),
-            Error::RangeTypeMissmatch { key_path, type1, type2 } => write!(f, "Conflicting range value type at key \"{}\", found type {} but also type {}.", key_path, type1, type2),
-            Error::InvalidKey(key) => write!(f, "invalid key {:?}, it can't be used as a rust identifier, try removing whitespaces and special characters.", key),
+            Error::RangeTypeMissmatch {
+                key_path,
+                type1,
+                type2,
+            } => write!(
+                f,
+                "Conflicting range value type at key \"{key_path}\", found type {type1} but also type {type2}."
+            ),
+            Error::InvalidKey(key) => write!(
+                f,
+                "invalid key {key:?}, it can't be used as a rust identifier, try removing whitespaces and special characters."
+            ),
             Error::EmptyRange => write!(f, "empty ranges are not allowed"),
-            Error::InvalidRangeType(t) => write!(f, "invalid range type {:?}", t),
+            Error::InvalidRangeType(t) => write!(f, "invalid range type {t:?}"),
             Error::NestedRanges => write!(f, "nested ranges are not allowed"),
             Error::InvalidFallback => write!(f, "fallbacks are only allowed in last position"),
             Error::MultipleFallbacks => write!(f, "only one fallback is allowed"),
-            Error::MissingFallback(t) => write!(f, "range type {} require a fallback (or a fullrange \"..\")", t),
+            Error::MissingFallback(t) => write!(
+                f,
+                "range type {t} require a fallback (or a fullrange \"..\")"
+            ),
             Error::RangeSubkeys => write!(f, "subkeys for ranges are not allowed"),
             Error::SubKeyMissmatch { locale, key_path } => {
-                write!(f, "Missmatch value type beetween locale {:?} and default at key \"{}\": one has subkeys and the other has direct value.", locale, key_path)
-            },
-            Error::RangeNumberType { found, expected } => write!(f, "number type {} can't be used for range type {}", found, expected),
-            Error::ExplicitDefaultInDefault(key_path) => write!(f, "Explicit defaults (null) are not allowed in default locale, at key \"{}\"", key_path),
-            Error::RecursiveForeignKey { locale, key_path } => write!(f, "Borrow Error while linking foreign key at key \"{}\" in locale {:?}, check for recursive foreign key.", key_path, locale),
-            Error::MissingForeignKey { foreign_key, locale, key_path } => write!(f, "Invalid foreign key \"{}\" at key \"{}\" in locale {:?}, key don't exist.", foreign_key, key_path, locale),
-            Error::InvalidForeignKey { foreign_key, locale, key_path } => write!(f, "Invalid foreign key \"{}\" at key \"{}\" in locale {:?}, foreign key to subkeys are not allowed.", foreign_key, key_path, locale),
-            Error::UnknownFormatter { name, locale, key_path } => write!(f, "Unknown formatter {:?} at key \"{}\" in locale {:?}.", name, key_path, locale),
-            Error::ConflictingPluralRuleType { locale, key_path } => write!(f, "Found both ordinal and cardinal plurals for key \"{}\" in locale {:?}.", key_path, locale),
-            Error::InvalidForeignKeyArgs { locale, key_path, err } => write!(f, "Malformed foreign key args in locale {:?} at key \"{}\": {}.", locale, key_path, err),
-            Error::InvalidCountArg { locale, key_path, foreign_key } => write!(f, "Invalid arg \"count\" in locale {:?} at key \"{}\" to foreign key \"{}\": argument \"count\" for plurals or ranges can only be a literal number or a single variable.", locale, key_path, foreign_key),
-            Error::InvalidCountArgType { locale, key_path, foreign_key, input_type, range_type } => write!(f, "Invalid arg \"count\" in locale {:?} at key \"{}\" to foreign key \"{}\": argument \"count\" of type {} for range of type {} is not allowed.", locale, key_path, foreign_key, input_type, range_type),
-            Error::CountArgOutsideRange { locale, key_path, foreign_key, err } => write!(f, "Invalid arg \"count\" in locale {:?} at key \"{}\" to foreign key \"{}\": argument \"count\" is outside range: {}", locale, key_path, foreign_key, err),
-            Error::UnexpectedToken { locale, key_path, message } => write!(f, "Unexpected error occured while parsing key \"{}\" in locale {:?}: {}", key_path, locale, message),
-            Error::RangeAndPluralsMix { key_path } => write!(f, "mixing plurals and ranges are not supported yet, for key \"{}\"", key_path),
-            Error::PluralsAtNormalKey { key_path, locale } => write!(f, "In locale {:?} at key \"{}\", Found plurals but a key of that name is already present.", locale, key_path),
-            Error::DisabledFormatter { locale, key_path, formatter } => write!(f, "{}, at key \"{}\" in locale {:?}", formatter.err_message(), key_path, locale),
-            Error::DisabledPlurals { locale, key_path } => write!(f, "Plurals are not enabled, enable the \"plurals\" feature to use them, at key \"{}\" in locale {:?}", key_path, locale),
-            Error::NoFileFormats => write!(f, "No file formats has been provided for leptos_i18n. Supported formats are: json, json5 and yaml."),
-            Error::MultipleFilesFormats => write!(f, "Multiple file formats have been provided for leptos_i18n, choose only one. Supported formats are: json, json5 and yaml."),
-            Error::InvalidLocale {
-                locale,
-                err
-            } => write!(f, "Found invalid locale {:?}: {}", locale, err),
-            Error::PluralRulesError(plurals_error) => write!(f, "Error while computing plurals categories: {}", plurals_error),
+                write!(
+                    f,
+                    "Missmatch value type beetween locale {locale:?} and default at key \"{key_path}\": one has subkeys and the other has direct value."
+                )
+            }
+            Error::RangeNumberType { found, expected } => write!(
+                f,
+                "number type {found} can't be used for range type {expected}"
+            ),
+            Error::ExplicitDefaultInDefault(key_path) => write!(
+                f,
+                "Explicit defaults (null) are not allowed in default locale, at key \"{key_path}\""
+            ),
+            Error::RecursiveForeignKey { loc } => write!(
+                f,
+                "Borrow Error while linking foreign key at {loc}, check for recursive foreign key."
+            ),
+            Error::MissingForeignKey { foreign_key, loc } => write!(
+                f,
+                "Invalid foreign key \"{foreign_key}\" at {loc}, key don't exist."
+            ),
+            Error::InvalidForeignKey { foreign_key, loc } => write!(
+                f,
+                "Invalid foreign key \"{foreign_key}\" at {loc}, foreign key to subkeys are not allowed."
+            ),
+            Error::UnknownFormatter { name, loc } => {
+                write!(f, "Unknown formatter {name:?} at {loc}.")
+            }
+            Error::ConflictingPluralRuleType { loc } => {
+                write!(f, "Found both ordinal and cardinal plurals at {loc}.")
+            }
+            Error::InvalidForeignKeyArgs { loc, err } => {
+                write!(f, "Malformed foreign key args at {loc}: {err}.")
+            }
+            Error::InvalidCountArg { loc, foreign_key } => write!(
+                f,
+                "Invalid arg \"count\" at {loc} to foreign key \"{foreign_key}\": argument \"count\" for plurals or ranges can only be a literal number or a single variable."
+            ),
+            Error::InvalidCountArgType {
+                loc,
+                foreign_key,
+                input_type,
+                range_type,
+            } => write!(
+                f,
+                "Invalid arg \"count\" at {loc} to foreign key \"{foreign_key}\": argument \"count\" of type {input_type} for range of type {range_type} is not allowed."
+            ),
+            Error::CountArgOutsideRange {
+                loc,
+                foreign_key,
+                err,
+            } => write!(
+                f,
+                "Invalid arg \"count\" at {loc} to foreign key \"{foreign_key}\": argument \"count\" is outside range: {err}"
+            ),
+            Error::UnexpectedToken { loc, message } => write!(
+                f,
+                "Unexpected error occured while parsing at {loc}: {message}"
+            ),
+            Error::RangeAndPluralsMix { key_path } => write!(
+                f,
+                "mixing plurals and ranges are not supported yet, for key \"{key_path}\""
+            ),
+            Error::PluralsAtNormalKey { loc } => write!(
+                f,
+                "At {loc}, Found plurals but a key of that name is already present."
+            ),
+            Error::DisabledFormatter { loc, formatter_err } => {
+                write!(f, "{}, at {loc}", formatter_err)
+            }
+            Error::DisabledPlurals { loc } => write!(
+                f,
+                "Plurals are not enabled, enable the \"plurals\" feature to use them, at {loc}"
+            ),
+            Error::NoFileFormats => write!(
+                f,
+                "No file formats has been provided for leptos_i18n. Supported formats are: json, json5, yaml and toml."
+            ),
+            Error::MultipleFilesFormats => write!(
+                f,
+                "Multiple file formats have been provided for leptos_i18n, choose only one. Supported formats are: json, json5, yaml and toml."
+            ),
+            Error::InvalidLocale { locale, err } => {
+                write!(f, "Found invalid locale {locale:?}: {err}")
+            }
+            Error::PluralRulesError(plurals_error) => write!(
+                f,
+                "Error while computing plurals categories: {plurals_error}"
+            ),
             Error::MissingTranslationsURI => {
-                write!(f, "{:?} config option is missing. You are using dynamic loading in CSR, that value is required.", cfg_file::Field::TRANSLATIONS_URI)
-            },
+                write!(
+                    f,
+                    "{:?} config option is missing. You are using dynamic loading in CSR, that value is required.",
+                    cfg_file::Field::TRANSLATIONS_URI
+                )
+            }
+            Error::Custom(err) => {
+                write!(f, "{err}")
+            }
+            Error::InvalidFormatterArgName { loc, name, err } => write!(
+                f,
+                "Formatter argument name {name:?} is invalid at {loc}: {err}"
+            ),
+            Error::InvalidFormatterArg {
+                loc,
+                arg_name,
+                arg,
+                err,
+            } => write!(
+                f,
+                "Formatter argument value {arg:?} for argument name {arg_name:?} is invalid at {loc}: {err}"
+            ),
+            Error::InvalidFormatter { loc, err } => {
+                write!(f, "Formatter is invalid at {loc}: {err}")
+            }
+            Error::InvalidAttribute {
+                loc,
+                attr_name,
+                attr_value,
+                err,
+            } => write!(
+                f,
+                "Invalid component attribute value {attr_value:?} for attribute {attr_name:?} at {loc}: {err}"
+            ),
+            Error::InvalidForeignKeyArgForAttribute {
+                loc,
+                arg_name,
+                foreign_key,
+            } => write!(
+                f,
+                "Invalid foreign key argument {arg_name:?} to key \"{foreign_key}\" at {loc}: argument to attributes must be either a variable or a literal (boolean, string, numbers)"
+            ),
+            Error::InvalidAttributeName { loc, value } => {
+                write!(f, "Invalid attribute name {value:?} at {loc}")
+            }
+            Error::UnknownLocaleInInherit { loc, locale } => {
+                write!(
+                    f,
+                    "Tried to declare inheritance for an unknown locale \"{locale}\" at {loc}, make sure to add it before declaring the inheritance."
+                )
+            }
+            Error::DefaultLocaleCantInherit { loc } => {
+                write!(
+                    f,
+                    "Tried to declare inheritance for the default locale at {loc}"
+                )
+            }
         }
     }
 }
 
-pub type Result<T, E = Box<Error>> = core::result::Result<T, E>;
+impl Error {
+    pub fn custom(err: impl ToString) -> Self {
+        Self::Custom(err.to_string())
+    }
+}
+
+impl From<io::Error> for Error {
+    fn from(value: io::Error) -> Self {
+        Self::IoError(value)
+    }
+}
+
+pub struct BoxedError(Box<Error>);
+
+impl<T: Into<Error>> From<T> for BoxedError {
+    fn from(value: T) -> Self {
+        BoxedError(Box::new(value.into()))
+    }
+}
+
+impl From<BoxedError> for Box<dyn core::error::Error> {
+    fn from(value: BoxedError) -> Self {
+        value.0
+    }
+}
+
+impl From<Box<Error>> for BoxedError {
+    fn from(value: Box<Error>) -> Self {
+        BoxedError(value)
+    }
+}
+
+impl BoxedError {
+    pub fn into_inner(self) -> Error {
+        *self.0
+    }
+
+    pub fn into_boxed(self) -> Box<Error> {
+        self.0
+    }
+}
+
+impl Debug for BoxedError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        <Error as Debug>::fmt(&self.0, f)
+    }
+}
+
+impl Display for BoxedError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        <Error as Display>::fmt(&self.0, f)
+    }
+}
+
+pub type Result<T, E = BoxedError> = core::result::Result<T, E>;
 
 impl std::error::Error for Error {}
+
+use super::plurals::{PluralForm, PluralRuleType};
+
+#[derive(Debug)]
+pub enum Warning {
+    MissingKey {
+        loc: Location,
+    },
+    SurplusKey {
+        loc: Location,
+    },
+    UnusedForm {
+        loc: Location,
+        form: PluralForm,
+        rule_type: PluralRuleType,
+    },
+    NonUnicodePath {
+        locale: Key,
+        namespace: Option<Key>,
+        path: std::path::PathBuf,
+    },
+    UnexpectedCharsAfterFormatter {
+        loc: Location,
+        formatter_name: String,
+        chars: String,
+    },
+    Custom(String),
+}
+
+impl Warning {
+    pub fn custom(err: impl ToString) -> Self {
+        Warning::Custom(err.to_string())
+    }
+}
+
+impl Display for Warning {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Warning::MissingKey { loc } => {
+                write!(
+                    f,
+                    "Missing key \"{}\" in locale {:?}",
+                    loc.key_path, loc.locale
+                )
+            }
+            Warning::SurplusKey { loc } => write!(
+                f,
+                "Key \"{}\" is present in locale {:?} but not in default locale, it is ignored",
+                loc.key_path, loc.locale
+            ),
+            Warning::UnusedForm {
+                loc,
+                form,
+                rule_type,
+            } => {
+                write!(
+                    f,
+                    "At key \"{}\", locale {:?} does not use {rule_type} plural form \"{form}\", it is still kept but is useless.",
+                    loc.key_path, loc.locale
+                )
+            }
+            Warning::NonUnicodePath {
+                locale,
+                namespace: None,
+                path,
+            } => write!(
+                f,
+                "File path for locale {locale:?} is not valid UTF8, can't add it to build script depedencies. Path: {path:?}"
+            ),
+            Warning::NonUnicodePath {
+                locale,
+                namespace: Some(ns),
+                path,
+            } => write!(
+                f,
+                "File path for locale {locale:?} in namespace {ns:?} is not valid UTF8, can't add it to build script depedencies. Path: {path:?}"
+            ),
+            Warning::Custom(warn) => write!(f, "{warn}"),
+            Warning::UnexpectedCharsAfterFormatter {
+                loc,
+                chars,
+                formatter_name,
+            } => write!(
+                f,
+                "Unexpected characters {chars:?} after formatter {formatter_name:?} at {loc}"
+            ),
+        }
+    }
+}
+
+#[derive(Default)]
+pub struct Diagnostics {
+    errors: RefCell<Vec<Error>>,
+    warnings: RefCell<Vec<Warning>>,
+    has_ranges: Cell<bool>,
+}
+
+impl Diagnostics {
+    pub fn new() -> Self {
+        Default::default()
+    }
+
+    pub fn emit_error(&self, error: Error) {
+        self.errors.borrow_mut().push(error);
+    }
+
+    pub fn emit_custom_error(&self, err: impl ToString) {
+        self.emit_error(Error::custom(err));
+    }
+
+    pub fn emit_custom_warning(&self, err: impl ToString) {
+        self.emit_warning(Warning::custom(err));
+    }
+
+    pub fn emit_warning(&self, warning: Warning) {
+        self.warnings.borrow_mut().push(warning);
+    }
+
+    pub fn errors(&self) -> Ref<'_, [Error]> {
+        let errors = self.errors.borrow();
+        Ref::map(errors, Vec::as_slice)
+    }
+
+    pub fn warnings(&self) -> Ref<'_, [Warning]> {
+        let warnings = self.warnings.borrow();
+        Ref::map(warnings, Vec::as_slice)
+    }
+
+    pub fn borrow(&self) -> (Ref<'_, [Error]>, Ref<'_, [Warning]>) {
+        (self.errors(), self.warnings())
+    }
+
+    pub fn has_ranges(&self) -> bool {
+        self.has_ranges.get()
+    }
+
+    pub fn set_has_ranges(&self) {
+        self.has_ranges.set(true);
+    }
+}
+
+impl ToTokens for Diagnostics {
+    fn to_token_stream(&self) -> proc_macro2::TokenStream {
+        let (errors, warnings) = self.borrow();
+        let iter = errors.iter().map(ToString::to_string);
+        let warnings = generate_warnings(&warnings);
+
+        quote! {
+            #(
+                compile_error!(#iter);
+            )*
+
+            #warnings
+        }
+    }
+    fn to_tokens(&self, tokens: &mut proc_macro2::TokenStream) {
+        let ts = Self::to_token_stream(self);
+        tokens.extend(ts);
+    }
+}
+
+fn warning_fn((index, warning): (usize, &Warning)) -> TokenStream {
+    let msg = warning.to_string();
+    let fn_name = format_ident!("w{}", index);
+    quote! {
+        #[deprecated(note = #msg)]
+        fn #fn_name() {
+            unimplemented!()
+        }
+    }
+}
+
+fn generate_warnings_inner(warnings: &[Warning]) -> TokenStream {
+    let warning_fns = warnings.iter().enumerate().map(warning_fn);
+
+    let fn_calls = (0..warnings.len()).map(|i| {
+        let fn_name = format_ident!("w{}", i);
+        quote!(#fn_name();)
+    });
+
+    quote! {
+        #[allow(unused)]
+        fn warnings() {
+            #(
+                #warning_fns
+            )*
+
+            #(
+                #fn_calls
+            )*
+        }
+    }
+}
+
+pub fn generate_warnings(warnings: &[Warning]) -> Option<TokenStream> {
+    if warnings.is_empty() {
+        None
+    } else {
+        Some(generate_warnings_inner(warnings))
+    }
+}
