@@ -1,15 +1,25 @@
-use leptos_i18n_build::TranslationsInfos;
+use leptos_i18n_build::{Config, TranslationsInfos};
+use std::error::Error;
 use std::path::PathBuf;
 
-fn main() {
-    println!("cargo:rerun-if-changed=build.rs");
-    println!("cargo:rerun-if-changed=Cargo.toml");
+fn main() -> Result<(), Box<dyn Error>> {
+    println!("cargo::rerun-if-changed=build.rs");
+    println!("cargo::rerun-if-changed=Cargo.toml");
 
     let mod_directory = PathBuf::from(std::env::var_os("OUT_DIR").unwrap()).join("baked_data");
+    let i18n_mod_directory = PathBuf::from(std::env::var_os("OUT_DIR").unwrap()).join("i18n");
 
-    let translations_infos = TranslationsInfos::parse().unwrap();
+    let cfg = Config::new("en")?.add_locale("fr")?;
+
+    let translations_infos = TranslationsInfos::parse(cfg)?;
+
+    translations_infos.emit_diagnostics();
 
     translations_infos.rerun_if_locales_changed();
 
-    translations_infos.generate_data(mod_directory).unwrap();
+    translations_infos.generate_data(mod_directory)?;
+
+    translations_infos.generate_i18n_module(i18n_mod_directory)?;
+
+    Ok(())
 }
