@@ -1,100 +1,123 @@
 use std::fmt::{self, Display};
 
-use icu_calendar::AnyCalendar;
-use icu_datetime::{input::DateInput, options::length};
+use icu_calendar::{AnyCalendar, Date, Ref};
+use icu_datetime::{
+    options::{Alignment, Length, YearStyle},
+    scaffold::ConvertCalendar,
+};
 use leptos::IntoView;
 
 use crate::Locale;
 
-/// Marker trait for types that can lend a reference to
-/// `T: icu::datetime::input::DateInput<Calendar = icu::calendar::AnyCalendar>`.
+/// Provides a reference conversion of a type into an ICU4X DateTime, allowing non-destructive access to calendar-specific date operations.
 pub trait AsIcuDate {
-    /// The returned `T: DateInput<Calendar = AnyCalendar>`.
-    type Date: DateInput<Calendar = AnyCalendar>;
-
-    /// Lend a reference to `Self::Date`.
-    fn as_icu_date(&self) -> &Self::Date;
+    /// The associated Date type that represents a calendar-specific date value in ICU4X format.
+    type Date<'a>: ConvertCalendar<Converted<'a> = Date<Ref<'a, AnyCalendar>>>;
+    /// Returns a reference to the calendar-specific Date representation of this value.
+    fn as_icu_date<'a>(&self) -> &Self::Date<'a>;
 }
 
-impl<T: DateInput<Calendar = AnyCalendar>> AsIcuDate for T {
-    type Date = Self;
-
-    fn as_icu_date(&self) -> &Self::Date {
+impl<T> AsIcuDate for T
+where
+    T: for<'a> ConvertCalendar<Converted<'a> = Date<Ref<'a, AnyCalendar>>>,
+{
+    type Date<'a> = Self;
+    fn as_icu_date<'a>(&self) -> &Self::Date<'a> {
         self
     }
 }
 
-/// Marker trait for types that can be turned into a type
-/// `T: icu::datetime::input::DateInput<Calendar = icu::calendar::AnyCalendar>`.
+/// Enables consuming conversion of a type into an ICU4X Date, transforming the original value into a calendar-specific date representation.
 pub trait IntoIcuDate {
-    /// The returned `T: DateInput<Calendar = AnyCalendar>`.
-    type Date: DateInput<Calendar = AnyCalendar>;
-
-    /// Consume self and return a `T: DateInput<Calendar = AnyCalendar>`.
-    fn into_icu_date(self) -> Self::Date;
+    /// The associated Date type that represents a calendar-specific date value in ICU4X format.
+    type Date<'a>: ConvertCalendar<Converted<'a> = Date<Ref<'a, AnyCalendar>>>;
+    /// Consumes self to produce its calendar-specific Date representation.
+    fn into_icu_date<'a>(self) -> Self::Date<'a>;
 }
 
-impl<T: DateInput<Calendar = AnyCalendar>> IntoIcuDate for T {
-    type Date = Self;
+impl<T> IntoIcuDate for T
+where
+    T: for<'a> ConvertCalendar<Converted<'a> = Date<Ref<'a, AnyCalendar>>>,
+{
+    type Date<'a> = Self;
 
-    fn into_icu_date(self) -> Self::Date {
+    fn into_icu_date<'a>(self) -> Self::Date<'a> {
         self
     }
 }
 
-/// Marker trait for types that produce a `T: DateInput<Calendar = AnyCalendar>`.
+/// Defines a stateless, thread-safe function type that can repeatedly generate ICU4X Date instances for formatting operations.
 pub trait DateFormatterInputFn: 'static + Clone + Send + Sync {
-    /// The returned `T: DateInput<Calendar = AnyCalendar>`.
-    type Date: DateInput<Calendar = AnyCalendar>;
+    /// The associated Date type that represents a calendar-specific date value in ICU4X format.
+    type Date<'a>: ConvertCalendar<Converted<'a> = Date<Ref<'a, AnyCalendar>>>;
 
-    /// Produce a `Self::Date`.
-    fn to_icu_date(&self) -> Self::Date;
+    /// Generates a new calendar-specific Date instance for formatting operations.
+    fn to_icu_date<'a>(&self) -> Self::Date<'a>;
 }
 
-impl<T: IntoIcuDate, F: Fn() -> T + Clone + Send + Sync + 'static> DateFormatterInputFn for F {
-    type Date = T::Date;
+impl<T, F> DateFormatterInputFn for F
+where
+    T: IntoIcuDate,
+    F: Fn() -> T + Clone + Send + Sync + 'static,
+{
+    type Date<'a> = T::Date<'a>;
 
-    fn to_icu_date(&self) -> Self::Date {
+    fn to_icu_date<'a>(&self) -> Self::Date<'a> {
         IntoIcuDate::into_icu_date(self())
     }
 }
 
 #[doc(hidden)]
-pub fn format_date_to_view<L: Locale>(
+pub fn format_date_to_view<L, I>(
     locale: L,
-    date: impl DateFormatterInputFn,
-    length: length::Date,
-) -> impl IntoView + Clone {
-    let date_formatter = super::get_date_formatter(locale, length);
+    date: I,
+    length: Length,
+    alignment: Alignment,
+    year_style: YearStyle,
+) -> impl IntoView + Clone
+where
+    L: Locale,
+    I: DateFormatterInputFn,
+{
+    let date_formatter = super::get_date_formatter(locale, length, alignment, year_style);
 
     move || {
         let date = date.to_icu_date();
-        date_formatter
-            .format_to_string(&date)
-            .expect("The date formatter to return a formatted date.")
+
+        date_formatter.format(&date).to_string()
     }
 }
 
 #[doc(hidden)]
-pub fn format_date_to_formatter<L: Locale>(
+pub fn format_date_to_formatter<L, I>(
     f: &mut fmt::Formatter<'_>,
     locale: L,
-    date: &impl AsIcuDate,
-    length: length::Date,
-) -> fmt::Result {
-    let formatted_date = format_date_to_display(locale, date, length);
+    date: &I,
+    length: Length,
+    alignment: Alignment,
+    year_style: YearStyle,
+) -> fmt::Result
+where
+    L: Locale,
+    I: AsIcuDate,
+{
+    let formatted_date = format_date_to_display(locale, date, length, alignment, year_style);
     Display::fmt(&formatted_date, f)
 }
 
 #[doc(hidden)]
-pub fn format_date_to_display<L: Locale>(
+pub fn format_date_to_display<L, I>(
     locale: L,
-    date: &impl AsIcuDate,
-    length: length::Date,
-) -> impl Display {
-    let date_formatter = super::get_date_formatter(locale, length);
+    date: &I,
+    length: Length,
+    alignment: Alignment,
+    year_style: YearStyle,
+) -> impl Display
+where
+    L: Locale,
+    I: AsIcuDate,
+{
+    let date_formatter = super::get_date_formatter(locale, length, alignment, year_style);
     let date = date.as_icu_date();
-    date_formatter
-        .format(date)
-        .expect("The date formatter to return a formatted date.")
+    date_formatter.format(date)
 }
