@@ -21,9 +21,12 @@ pub use date::*;
 #[cfg(feature = "format_datetime")]
 pub use datetime::*;
 #[cfg(feature = "format_datetime")]
-use icu_datetime::{options::length, DateFormatter, DateTimeFormatter, TimeFormatter};
+use icu_datetime::{
+    DateTimeFormatter, DateTimeFormatterLoadError, NoCalendarFormatter, fieldsets,
+    options::{Alignment, Length, TimePrecision, YearStyle},
+};
 #[cfg(feature = "format_list")]
-use icu_list::{ListFormatter, ListLength};
+use icu_list::{ListFormatter, options::ListLength};
 #[cfg(feature = "plurals")]
 use icu_plurals::{PluralRuleType, PluralRules};
 #[cfg(feature = "format_list")]
@@ -42,11 +45,11 @@ pub use time::*;
 ))]
 use crate::Locale;
 #[cfg(feature = "format_nums")]
-use icu_decimal::options::FixedDecimalFormatterOptions;
+use icu_decimal::DecimalFormatter;
+#[cfg(feature = "format_nums")]
+use icu_decimal::options::DecimalFormatterOptions;
 #[cfg(feature = "format_nums")]
 use icu_decimal::options::GroupingStrategy;
-#[cfg(feature = "format_nums")]
-use icu_decimal::FixedDecimalFormatter;
 #[cfg(feature = "format_currency")]
 use icu_experimental::dimension::currency::formatter::CurrencyFormatter;
 #[cfg(feature = "format_currency")]
@@ -72,7 +75,7 @@ fn get_currency_formatter<L: Locale>(
         let currency_formatter = currency_formatters.entry(width.into()).or_insert_with(|| {
             let formatter = formatters
                 .provider
-                .try_new_currency_formatter(&locale.into(), CurrencyFormatterOptions::from(width))
+                .try_new_currency_formatter(locale, CurrencyFormatterOptions::from(width))
                 .expect("A CurrencyFormatter");
             Box::leak(Box::new(formatter))
         });
@@ -84,7 +87,7 @@ fn get_currency_formatter<L: Locale>(
 fn get_num_formatter<L: Locale>(
     locale: L,
     grouping_strategy: GroupingStrategy,
-) -> &'static FixedDecimalFormatter {
+) -> &'static DecimalFormatter {
     use data_provider::IcuDataProvider;
 
     inner::FORMATTERS.with_mut(|formatters| {
@@ -93,11 +96,8 @@ fn get_num_formatter<L: Locale>(
         let num_formatter = num_formatters.entry(grouping_strategy).or_insert_with(|| {
             let formatter = formatters
                 .provider
-                .try_new_num_formatter(
-                    &locale.into(),
-                    FixedDecimalFormatterOptions::from(grouping_strategy),
-                )
-                .expect("A FixedDecimalFormatter");
+                .try_new_num_formatter(locale, DecimalFormatterOptions::from(grouping_strategy))
+                .expect("A DecimalFormatter");
             Box::leak(Box::new(formatter))
         });
         *num_formatter
@@ -105,37 +105,51 @@ fn get_num_formatter<L: Locale>(
 }
 
 #[cfg(feature = "format_datetime")]
-fn get_date_formatter<L: Locale>(locale: L, length: length::Date) -> &'static DateFormatter {
+fn get_date_formatter<L: Locale>(
+    locale: L,
+    length: Length,
+    alignment: Alignment,
+    year_style: YearStyle,
+) -> &'static DateTimeFormatter<fieldsets::YMD> {
     use data_provider::IcuDataProvider;
 
     inner::FORMATTERS.with_mut(|formatters| {
         let locale = locale.as_icu_locale();
         let date_formatters = formatters.date.entry(locale).or_default();
-        let date_formatter = date_formatters.entry(length).or_insert_with(|| {
-            let formatter = formatters
-                .provider
-                .try_new_date_formatter(&locale.into(), length)
-                .expect("A DateFormatter");
-            Box::leak(Box::new(formatter))
-        });
+        let date_formatter = date_formatters
+            .entry((length, alignment, year_style))
+            .or_insert_with(|| {
+                let formatter = formatters
+                    .provider
+                    .try_new_date_formatter(locale, length, alignment, year_style)
+                    .expect("A DateFormatter");
+                Box::leak(Box::new(formatter))
+            });
         *date_formatter
     })
 }
 
 #[cfg(feature = "format_datetime")]
-fn get_time_formatter<L: Locale>(locale: L, length: length::Time) -> &'static TimeFormatter {
+fn get_time_formatter<L: Locale>(
+    locale: L,
+    length: Length,
+    alignment: Alignment,
+    time_precision: TimePrecision,
+) -> &'static NoCalendarFormatter<fieldsets::T> {
     use data_provider::IcuDataProvider;
 
     inner::FORMATTERS.with_mut(|formatters| {
         let locale = locale.as_icu_locale();
         let time_formatters = formatters.time.entry(locale).or_default();
-        let time_formatter = time_formatters.entry(length).or_insert_with(|| {
-            let formatter = formatters
-                .provider
-                .try_new_time_formatter(&locale.into(), length)
-                .expect("A TimeFormatter");
-            Box::leak(Box::new(formatter))
-        });
+        let time_formatter = time_formatters
+            .entry((length, alignment, time_precision))
+            .or_insert_with(|| {
+                let formatter = formatters
+                    .provider
+                    .try_new_time_formatter(locale, length, alignment, time_precision)
+                    .expect("A TimeFormatter");
+                Box::leak(Box::new(formatter))
+            });
         *time_formatter
     })
 }
@@ -143,21 +157,28 @@ fn get_time_formatter<L: Locale>(locale: L, length: length::Time) -> &'static Ti
 #[cfg(feature = "format_datetime")]
 fn get_datetime_formatter<L: Locale>(
     locale: L,
-    date_length: length::Date,
-    time_length: length::Time,
-) -> &'static DateTimeFormatter {
+    length: Length,
+    alignment: Alignment,
+    time_precision: TimePrecision,
+    year_style: YearStyle,
+) -> &'static DateTimeFormatter<fieldsets::YMDT> {
     use data_provider::IcuDataProvider;
 
     inner::FORMATTERS.with_mut(|formatters| {
         let locale = locale.as_icu_locale();
         let datetime_formatters = formatters.datetime.entry(locale).or_default();
         let datetime_formatter = datetime_formatters
-            .entry((date_length, time_length))
+            .entry((length, alignment, time_precision, year_style))
             .or_insert_with(|| {
-                let options = length::Bag::from_date_time_style(date_length, time_length);
                 let formatter = formatters
                     .provider
-                    .try_new_datetime_formatter(&locale.into(), options.into())
+                    .try_new_datetime_formatter(
+                        locale,
+                        length,
+                        alignment,
+                        time_precision,
+                        year_style,
+                    )
                     .expect("A DateTimeFormatter");
                 Box::leak(Box::new(formatter))
             });
@@ -198,7 +219,7 @@ pub fn get_plural_rules<L: Locale>(
         let plural_rules = plural_rules.entry(plural_rule_type).or_insert_with(|| {
             let plural_rules = formatters
                 .provider
-                .try_new_plural_rules(&locale.into(), plural_rule_type)
+                .try_new_plural_rules(locale, plural_rule_type)
                 .expect("A PluralRules");
             Box::leak(Box::new(plural_rules))
         });
@@ -215,12 +236,18 @@ pub fn get_plural_rules<L: Locale>(
 ))]
 pub(crate) mod inner {
     use super::*;
-    use icu_locid::Locale as IcuLocale;
+    use icu_locale::Locale as IcuLocale;
     use std::{
         collections::HashMap,
         sync::{OnceLock, RwLock},
     };
 
+    #[cfg(feature = "format_datetime")]
+    type DateTimeFormatterKey = (Length, Alignment, TimePrecision, YearStyle);
+    #[cfg(feature = "format_datetime")]
+    type DateFormatterKey = (Length, Alignment, YearStyle);
+    #[cfg(feature = "format_datetime")]
+    type TimeFormatterKey = (Length, Alignment, TimePrecision);
     // Formatters cache
     //
     // The reason we leak the formatter is so that we can get a static ref,
@@ -254,16 +281,21 @@ pub(crate) mod inner {
             HashMap<super::currency::Width, &'static CurrencyFormatter>,
         >,
         #[cfg(feature = "format_nums")]
-        pub num:
-            HashMap<&'static IcuLocale, HashMap<GroupingStrategy, &'static FixedDecimalFormatter>>,
+        pub num: HashMap<&'static IcuLocale, HashMap<GroupingStrategy, &'static DecimalFormatter>>,
         #[cfg(feature = "format_datetime")]
-        pub date: HashMap<&'static IcuLocale, HashMap<length::Date, &'static DateFormatter>>,
+        pub date: HashMap<
+            &'static IcuLocale,
+            HashMap<DateFormatterKey, &'static DateTimeFormatter<fieldsets::YMD>>,
+        >,
         #[cfg(feature = "format_datetime")]
-        pub time: HashMap<&'static IcuLocale, HashMap<length::Time, &'static TimeFormatter>>,
+        pub time: HashMap<
+            &'static IcuLocale,
+            HashMap<TimeFormatterKey, &'static NoCalendarFormatter<fieldsets::T>>,
+        >,
         #[cfg(feature = "format_datetime")]
         pub datetime: HashMap<
             &'static IcuLocale,
-            HashMap<(length::Date, length::Time), &'static DateTimeFormatter>,
+            HashMap<DateTimeFormatterKey, &'static DateTimeFormatter<fieldsets::YMDT>>,
         >,
         #[cfg(feature = "format_list")]
         pub list: HashMap<
@@ -315,89 +347,105 @@ pub(crate) mod data_provider {
 
     #[cfg(any(
         feature = "format_nums",
+        feature = "format_list",
+        feature = "plurals",
+        feature = "format_currency",
+    ))]
+    use icu_provider::DataError;
+
+    #[cfg(any(
+        feature = "format_nums",
         feature = "format_datetime",
         feature = "format_list",
         feature = "plurals",
         feature = "format_currency",
     ))]
-    use icu_provider::DataLocale;
+    use icu_locale::Locale;
 
     /// Trait for custom ICU data providers.
     pub trait IcuDataProvider: Send + Sync + 'static {
-        /// Tries to create a new `FixedDecimalFormatter` with the given options
+        /// Tries to create a new `DecimalFormatter` with the given options
         #[cfg(feature = "format_nums")]
         fn try_new_num_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_decimal::options::FixedDecimalFormatterOptions,
-        ) -> Result<FixedDecimalFormatter, icu_decimal::DecimalError>;
+            locale: &Locale,
+            options: icu_decimal::options::DecimalFormatterOptions,
+        ) -> Result<DecimalFormatter, DataError>;
 
         /// Tries to create a new `DateFormatter` with the given options
         #[cfg(feature = "format_datetime")]
         fn try_new_date_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Date,
-        ) -> Result<DateFormatter, icu_datetime::DateTimeError>;
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMD>, DateTimeFormatterLoadError>;
 
         /// Tries to create a new `TimeFormatter` with the given options
         #[cfg(feature = "format_datetime")]
         fn try_new_time_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Time,
-        ) -> Result<TimeFormatter, icu_datetime::DateTimeError>;
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+        ) -> Result<NoCalendarFormatter<fieldsets::T>, DateTimeFormatterLoadError>;
 
         /// Tries to create a new `DateTimeFormatter` with the given options
         #[cfg(feature = "format_datetime")]
         fn try_new_datetime_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_datetime::options::DateTimeFormatterOptions,
-        ) -> Result<DateTimeFormatter, icu_datetime::DateTimeError>;
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMDT>, DateTimeFormatterLoadError>;
 
         /// Tries to create a and `ListFormatter` with the given options
         #[cfg(feature = "format_list")]
         fn try_new_and_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError>;
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError>;
 
         /// Tries to create a new or `ListFormatter` with the given options
         #[cfg(feature = "format_list")]
         fn try_new_or_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError>;
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError>;
 
         /// Tries to create a new unit `ListFormatter` with the given options
         #[cfg(feature = "format_list")]
         fn try_new_unit_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError>;
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError>;
 
         /// Tries to create a new `PluralRules` with the given options
         #[cfg(feature = "plurals")]
         fn try_new_plural_rules(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             rule_type: PluralRuleType,
-        ) -> Result<PluralRules, icu_plurals::PluralsError>;
+        ) -> Result<PluralRules, DataError>;
         ///
         /// Tries to create a new `CurrencyFormatter` with the given options
         #[cfg(feature = "format_currency")]
         fn try_new_currency_formatter(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             options: CurrencyFormatterOptions,
         ) -> Result<CurrencyFormatter, icu_provider::DataError>;
     }
 
     #[cfg(feature = "icu_compiled_data")]
+    #[allow(dead_code)]
     #[derive(Default)]
     pub struct BakedDataProvider;
 
@@ -406,82 +454,107 @@ pub(crate) mod data_provider {
         #[cfg(feature = "format_nums")]
         fn try_new_num_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_decimal::options::FixedDecimalFormatterOptions,
-        ) -> Result<FixedDecimalFormatter, icu_decimal::DecimalError> {
-            FixedDecimalFormatter::try_new(locale, options)
+            locale: &Locale,
+            options: icu_decimal::options::DecimalFormatterOptions,
+        ) -> Result<DecimalFormatter, DataError> {
+            DecimalFormatter::try_new(locale.into(), options)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_date_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Date,
-        ) -> Result<DateFormatter, icu_datetime::DateTimeError> {
-            DateFormatter::try_new_with_length(locale, length)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMD>, DateTimeFormatterLoadError> {
+            let fset = fieldsets::YMD::for_length(length)
+                .with_alignment(alignment)
+                .with_year_style(year_style);
+            DateTimeFormatter::try_new(locale.into(), fset)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_time_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Time,
-        ) -> Result<TimeFormatter, icu_datetime::DateTimeError> {
-            TimeFormatter::try_new_with_length(locale, length)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+        ) -> Result<NoCalendarFormatter<fieldsets::T>, DateTimeFormatterLoadError> {
+            let fset = fieldsets::T::for_length(length)
+                .with_alignment(alignment)
+                .with_time_precision(time_precision);
+            NoCalendarFormatter::try_new(locale.into(), fset)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_datetime_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_datetime::options::DateTimeFormatterOptions,
-        ) -> Result<DateTimeFormatter, icu_datetime::DateTimeError> {
-            DateTimeFormatter::try_new(locale, options)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMDT>, DateTimeFormatterLoadError> {
+            let fset = fieldsets::YMDT::for_length(length)
+                .with_alignment(alignment)
+                .with_time_precision(time_precision)
+                .with_year_style(year_style);
+            DateTimeFormatter::try_new(locale.into(), fset)
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_and_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
-            ListFormatter::try_new_and_with_length(locale, style)
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
+            use icu_list::options::ListFormatterOptions;
+            let options = ListFormatterOptions::default().with_length(length);
+            ListFormatter::try_new_and(locale.into(), options)
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_or_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
-            ListFormatter::try_new_or_with_length(locale, style)
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
+            use icu_list::options::ListFormatterOptions;
+            let options = ListFormatterOptions::default().with_length(length);
+            ListFormatter::try_new_or(locale.into(), options)
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_unit_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
-            ListFormatter::try_new_unit_with_length(locale, style)
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
+            use icu_list::options::ListFormatterOptions;
+            let options = ListFormatterOptions::default().with_length(length);
+            ListFormatter::try_new_unit(locale.into(), options)
         }
 
         #[cfg(feature = "plurals")]
         fn try_new_plural_rules(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             rule_type: PluralRuleType,
-        ) -> Result<PluralRules, icu_plurals::PluralsError> {
-            PluralRules::try_new(locale, rule_type)
+        ) -> Result<PluralRules, DataError> {
+            use icu_plurals::PluralRulesOptions;
+            let options = PluralRulesOptions::default().with_type(rule_type);
+            PluralRules::try_new(locale.into(), options)
         }
 
         #[cfg(feature = "format_currency")]
         fn try_new_currency_formatter(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             options: CurrencyFormatterOptions,
-        ) -> Result<CurrencyFormatter, icu_provider::DataError> {
-            CurrencyFormatter::try_new(locale, options)
+        ) -> Result<CurrencyFormatter, DataError> {
+            CurrencyFormatter::try_new(locale.into(), options)
         }
     }
 
@@ -502,84 +575,99 @@ pub(crate) mod data_provider {
         #[cfg(feature = "format_nums")]
         fn try_new_num_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_decimal::options::FixedDecimalFormatterOptions,
-        ) -> Result<FixedDecimalFormatter, icu_decimal::DecimalError> {
+            locale: &Locale,
+            options: icu_decimal::options::DecimalFormatterOptions,
+        ) -> Result<DecimalFormatter, DataError> {
             self.get_provider().try_new_num_formatter(locale, options)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_date_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Date,
-        ) -> Result<DateFormatter, icu_datetime::DateTimeError> {
-            self.get_provider().try_new_date_formatter(locale, length)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMD>, DateTimeFormatterLoadError> {
+            self.get_provider()
+                .try_new_date_formatter(locale, length, alignment, year_style)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_time_formatter(
             &self,
-            locale: &DataLocale,
-            length: length::Time,
-        ) -> Result<TimeFormatter, icu_datetime::DateTimeError> {
-            self.get_provider().try_new_time_formatter(locale, length)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+        ) -> Result<NoCalendarFormatter<fieldsets::T>, DateTimeFormatterLoadError> {
+            self.get_provider()
+                .try_new_time_formatter(locale, length, alignment, time_precision)
         }
 
         #[cfg(feature = "format_datetime")]
         fn try_new_datetime_formatter(
             &self,
-            locale: &DataLocale,
-            options: icu_datetime::options::DateTimeFormatterOptions,
-        ) -> Result<DateTimeFormatter, icu_datetime::DateTimeError> {
-            self.get_provider()
-                .try_new_datetime_formatter(locale, options)
+            locale: &Locale,
+            length: Length,
+            alignment: Alignment,
+            time_precision: TimePrecision,
+            year_style: YearStyle,
+        ) -> Result<DateTimeFormatter<fieldsets::YMDT>, DateTimeFormatterLoadError> {
+            self.get_provider().try_new_datetime_formatter(
+                locale,
+                length,
+                alignment,
+                time_precision,
+                year_style,
+            )
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_and_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
             self.get_provider()
-                .try_new_and_list_formatter(locale, style)
+                .try_new_and_list_formatter(locale, length)
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_or_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
-            self.get_provider().try_new_or_list_formatter(locale, style)
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
+            self.get_provider()
+                .try_new_or_list_formatter(locale, length)
         }
 
         #[cfg(feature = "format_list")]
         fn try_new_unit_list_formatter(
             &self,
-            locale: &DataLocale,
-            style: ListLength,
-        ) -> Result<ListFormatter, icu_list::ListError> {
+            locale: &Locale,
+            length: ListLength,
+        ) -> Result<ListFormatter, DataError> {
             self.get_provider()
-                .try_new_unit_list_formatter(locale, style)
+                .try_new_unit_list_formatter(locale, length)
         }
 
         #[cfg(feature = "plurals")]
         fn try_new_plural_rules(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             rule_type: PluralRuleType,
-        ) -> Result<PluralRules, icu_plurals::PluralsError> {
+        ) -> Result<PluralRules, DataError> {
             self.get_provider().try_new_plural_rules(locale, rule_type)
         }
 
         #[cfg(feature = "format_currency")]
         fn try_new_currency_formatter(
             &self,
-            locale: &DataLocale,
+            locale: &Locale,
             options: CurrencyFormatterOptions,
-        ) -> Result<CurrencyFormatter, icu_provider::DataError> {
+        ) -> Result<CurrencyFormatter, DataError> {
             self.get_provider()
                 .try_new_currency_formatter(locale, options)
         }
