@@ -1,23 +1,26 @@
 use std::{collections::BTreeMap, fmt::Display};
 
 use leptos_i18n_parser::{
+    formatters::Formatters,
     parse_locales::{
+        ForeignKeysPaths, ParsedLocales,
         cfg_file::ConfigFile,
+        error::Diagnostics,
         locale::{Locale, LocalesOrNamespaces},
+        make_builder_keys,
+        options::{Config, ParseOptions},
         parsed_value::ParsedValue,
         ranges::{
             ParseRanges, Range, RangeNumber, Ranges, RangesInner, TypeOrRange, UntypedRangesInner,
         },
-        warning::Warnings,
-        ForeignKeysPaths,
     },
-    utils::{Key, KeyPath},
+    utils::{Key, KeyPath, Loc, ParseContext},
 };
 use proc_macro2::Span;
 use quote::ToTokens;
 use syn::{
-    parse::ParseBuffer, parse_macro_input, punctuated::Punctuated, token::Comma, Ident, Lit,
-    LitStr, Token,
+    Ident, Lit, LitStr, Token, parse::ParseBuffer, parse_macro_input, punctuated::Punctuated,
+    token::Comma,
 };
 
 pub fn declare_locales(tokens: proc_macro::TokenStream) -> proc_macro::TokenStream {
@@ -28,17 +31,22 @@ pub fn declare_locales(tokens: proc_macro::TokenStream) -> proc_macro::TokenStre
         foreign_keys_paths,
         interpolate_display,
     } = parse_macro_input!(tokens as ParsedInput);
-    let warnings = Warnings::new();
+    let diag = Diagnostics::new();
 
-    let result = super::load_locales_inner(
-        &crate_path,
-        &cfg_file,
-        locales,
-        foreign_keys_paths,
-        warnings,
-        None,
-        interpolate_display,
-    );
+    let mut cfg: Config = cfg_file.into();
+
+    cfg.options = ParseOptions::default().interpolate_display(interpolate_display);
+
+    let builder_keys = make_builder_keys(locales, &cfg, foreign_keys_paths, &diag).unwrap();
+
+    let parsed_locales = ParsedLocales {
+        cfg,
+        builder_keys,
+        diag,
+        tracked_files: None,
+    };
+
+    let result = leptos_i18n_codegen::gen_code(&parsed_locales, Some(&crate_path), true, None);
     match result {
         Ok(ts) => ts.into(),
         Err(err) => {
@@ -79,8 +87,8 @@ fn parse_array<T: syn::parse::Parse>(
 
 fn parse_str_value(
     input: syn::parse::ParseStream,
-    key_path: &mut KeyPath,
-    locale: &Key,
+    loc: &Loc,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<Option<ParsedValue>> {
     if !input.peek(LitStr) {
@@ -88,9 +96,30 @@ fn parse_str_value(
     }
     let lit_str = input.parse::<LitStr>()?;
     let value = lit_str.value();
-    ParsedValue::new(&value, key_path, locale, foreign_keys_paths)
-        .map(Some)
-        .map_err(|_| syn::Error::new_spanned(lit_str, "unknown formatter."))
+
+    let diag = Diagnostics::new();
+
+    let ctx = ParseContext {
+        loc: *loc,
+        foreign_keys_paths,
+        formatters,
+        diag: &diag,
+        parse_fns: ParsedValue::DEFAULT_FNS,
+    };
+
+    match ParsedValue::new(&ctx, &value) {
+        Ok(pv) => {
+            if let Some(err) = diag.errors().first() {
+                return emit_err(lit_str, err);
+            }
+            if let Some(warn) = diag.warnings().first() {
+                // TODO: warn instead of error
+                return emit_err(lit_str, warn);
+            }
+            Ok(Some(pv))
+        }
+        Err(err) => emit_err(lit_str, err),
+    }
 }
 
 fn parse_map_values(
@@ -98,6 +127,7 @@ fn parse_map_values(
     name: &Key,
     key_path: &mut KeyPath,
     locale: &Key,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<Option<ParsedValue>> {
     fn inner(input: syn::parse::ParseStream) -> syn::Result<ParseBuffer> {
@@ -109,7 +139,7 @@ fn parse_map_values(
         return Ok(None);
     };
 
-    let keys = parse_block_inner(content, key_path, locale, foreign_keys_paths)?;
+    let keys = parse_block_inner(content, key_path, locale, formatters, foreign_keys_paths)?;
 
     Ok(Some(ParsedValue::Subkeys(Some(Locale {
         top_locale_name: locale.clone(),
@@ -123,6 +153,7 @@ fn parse_map_values(
 pub struct ParseRangeSeed<'a> {
     pub key_path: &'a mut KeyPath,
     pub locale: &'a Key,
+    pub formatters: &'a Formatters,
     pub foreign_keys_paths: &'a ForeignKeysPaths,
 }
 
@@ -163,8 +194,12 @@ fn parse_range_pair<T: RangeNumber>(
     let content;
     syn::bracketed!(content in input);
 
-    let Some(parsed_value) =
-        parse_str_value(&content, seed.key_path, seed.locale, foreign_keys_paths)?
+    let loc = Loc {
+        locale: seed.locale,
+        key_path: seed.key_path,
+    };
+
+    let Some(parsed_value) = parse_str_value(&content, &loc, seed.formatters, foreign_keys_paths)?
     else {
         return Err(content.error("only strings are accepted here."));
     };
@@ -252,30 +287,39 @@ fn parse_values(
     input: syn::parse::ParseStream,
     key_path: &mut KeyPath,
     locale: &Key,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<(Key, ParsedValue)> {
     let ident: Ident = input.parse()?;
     let key = Key::from_ident(ident);
-    key_path.push_key(key.clone());
+    let mut pushed_key = key_path.push_key(key.clone());
     input.parse::<Token![:]>()?;
-    if let Some(parsed_value) = parse_str_value(input, key_path, locale, foreign_keys_paths)? {
-        key_path.pop_key();
+    let loc = Loc {
+        locale,
+        key_path: &pushed_key,
+    };
+    if let Some(parsed_value) = parse_str_value(input, &loc, formatters, foreign_keys_paths)? {
         return Ok((key, parsed_value));
     }
-    if let Some(parsed_value) = parse_map_values(input, &key, key_path, locale, foreign_keys_paths)?
-    {
-        key_path.pop_key();
+    if let Some(parsed_value) = parse_map_values(
+        input,
+        &key,
+        &mut pushed_key,
+        locale,
+        formatters,
+        foreign_keys_paths,
+    )? {
         return Ok((key, parsed_value));
     }
 
     let seed = ParseRangeSeed {
-        key_path,
+        key_path: &mut pushed_key,
         locale,
+        formatters,
         foreign_keys_paths,
     };
 
     if let Some(parsed_value) = parse_ranges(input, seed, foreign_keys_paths)? {
-        key_path.pop_key();
         return Ok((key, parsed_value));
     }
 
@@ -286,11 +330,13 @@ fn parse_block_inner(
     content: ParseBuffer,
     key_path: &mut KeyPath,
     locale: &Key,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<BTreeMap<Key, ParsedValue>> {
     let mut values = BTreeMap::new();
     while !content.is_empty() {
-        let (key, value) = parse_values(&content, key_path, locale, foreign_keys_paths)?;
+        let (key, value) =
+            parse_values(&content, key_path, locale, formatters, foreign_keys_paths)?;
         values.insert(key, value);
         if !content.is_empty() {
             content.parse::<Comma>()?;
@@ -303,16 +349,18 @@ fn parse_block(
     input: syn::parse::ParseStream,
     key_path: &mut KeyPath,
     locale: &Key,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<BTreeMap<Key, ParsedValue>> {
     let content;
     syn::braced!(content in input);
-    parse_block_inner(content, key_path, locale, foreign_keys_paths)
+    parse_block_inner(content, key_path, locale, formatters, foreign_keys_paths)
 }
 
 fn parse_locale(
     input: syn::parse::ParseStream,
     locale_key: Key,
+    formatters: &Formatters,
     foreign_keys_paths: &ForeignKeysPaths,
 ) -> syn::Result<Locale> {
     let loc_name_ident: Ident = input.parse()?;
@@ -324,7 +372,13 @@ fn parse_locale(
 
     let mut key_path = KeyPath::new(None);
 
-    let keys = parse_block(input, &mut key_path, &locale_key, foreign_keys_paths)?;
+    let keys = parse_block(
+        input,
+        &mut key_path,
+        &locale_key,
+        formatters,
+        foreign_keys_paths,
+    )?;
 
     if !input.is_empty() {
         input.parse::<Comma>()?;
@@ -384,7 +438,7 @@ impl syn::parse::Parse for ParsedInput {
         match locales_iter.next() {
             None => return emit_err(loc_ident, "missing locales."),
             Some(l) if Key::new(&l.value()).as_ref() != Some(&default) => {
-                return emit_err(l, "first locale should be the same as the default.")
+                return emit_err(l, "first locale should be the same as the default.");
             }
             _ => {}
         }
@@ -396,11 +450,12 @@ impl syn::parse::Parse for ParsedInput {
         // loc: { .. }
 
         let foreign_keys_paths = ForeignKeysPaths::new();
+        let formatters = Formatters::new();
 
         let locales = locales_key
             .iter()
             .cloned()
-            .map(|k| parse_locale(input, k, &foreign_keys_paths))
+            .map(|k| parse_locale(input, k, &formatters, &foreign_keys_paths))
             .collect::<syn::Result<Vec<_>>>()?;
 
         if !input.is_empty() {
@@ -409,8 +464,6 @@ impl syn::parse::Parse for ParsedInput {
 
         let crate_path = crate_path
             .unwrap_or_else(|| syn::Path::from(syn::Ident::new("leptos_i18n", Span::call_site())));
-
-        let interpolate_display = interpolate_display || cfg!(feature = "interpolate_display");
 
         Ok(ParsedInput {
             cfg_file: ConfigFile {
