@@ -1,27 +1,31 @@
-use std::fmt::Display;
-
-use crate::utils::formatter::Formatter;
+use leptos_i18n_parser::formatters::{Formatters, VarBounds};
 use quote::ToTokens;
+use std::fmt::Display;
 use syn::{
+    Expr, Ident, Token,
     parse::{ParseBuffer, ParseStream},
     token::Comma,
-    Expr, Ident, Token,
 };
 
 pub struct ParsedInput {
     pub context: Expr,
     pub value: Expr,
-    pub formatter: Formatter,
+    pub formatter: VarBounds,
 }
 
 fn emit_err<A, T: ToTokens, U: Display>(tokens: T, message: U) -> syn::Result<A> {
     Err(syn::Error::new_spanned(tokens, message))
 }
 
-fn parse_arg(input: ParseStream) -> syn::Result<(Ident, Ident)> {
+fn parse_arg(input: ParseStream) -> syn::Result<(Ident, Option<Ident>)> {
     let arg_name = input.parse::<Ident>()?;
-    input.parse::<Token![:]>()?;
-    let arg_value = input.parse::<Ident>()?;
+    let arg_value = if input.peek(Token![:]) {
+        input.parse::<Token![:]>()?;
+        let arg_value = input.parse::<Ident>()?;
+        Some(arg_value)
+    } else {
+        None
+    };
     Ok((arg_name, arg_value))
 }
 
@@ -31,35 +35,13 @@ fn is_parenthesized(input: ParseStream) -> syn::Result<ParseBuffer> {
     Ok(content)
 }
 
-fn convert_formatter_result(
-    res: Result<
-        Option<leptos_i18n_parser::utils::formatter::Formatter>,
-        leptos_i18n_parser::utils::formatter::Formatter,
-    >,
-    span: proc_macro2::Span,
-    err: syn::Error,
-) -> syn::Result<Formatter> {
-    match res {
-        Ok(Some(formatter)) => Ok(formatter.into()),
-        Ok(None) => Err(err),
-        Err(formatter) => Err(syn::Error::new(span, formatter.err_message())),
-    }
-}
-
 fn parse_formatter(
     input: ParseBuffer,
     formatter_name: Ident,
-    err: syn::Error,
-) -> syn::Result<Formatter> {
+    formatters: &Formatters,
+) -> syn::Result<VarBounds> {
     let args = input.parse_terminated(parse_arg, Token![;])?;
-    let args: Vec<_> = args.into_iter().collect();
-
-    let span = formatter_name.span();
-    let res = leptos_i18n_parser::utils::formatter::Formatter::from_name_and_args(
-        formatter_name,
-        Some(&args),
-    );
-    convert_formatter_result(res, span, err)
+    formatters.parse_from_tt(formatter_name, Some(args))
 }
 
 impl syn::parse::Parse for ParsedInput {
@@ -74,16 +56,11 @@ impl syn::parse::Parse for ParsedInput {
         }
         input.parse::<Token![:]>()?;
         let formatter_name = input.parse::<Ident>()?;
-        let formatter_name_err = syn::Error::new(formatter_name.span(), "unknown formatter name.");
+        let formatters = Formatters::new();
         let formatter = if let Ok(args) = is_parenthesized(input) {
-            parse_formatter(args, formatter_name, formatter_name_err)?
+            parse_formatter(args, formatter_name, &formatters)?
         } else {
-            let span = formatter_name.span();
-            let res = leptos_i18n_parser::utils::formatter::Formatter::from_name_and_args(
-                formatter_name,
-                None,
-            );
-            convert_formatter_result(res, span, formatter_name_err)?
+            formatters.parse_from_tt(formatter_name, None)?
         };
 
         Ok(ParsedInput {
