@@ -1,9 +1,9 @@
 use proc_macro2::Span;
 use quote::ToTokens;
 use std::{collections::BTreeMap, fmt::Display};
-use syn::{parse::ParseBuffer, spanned::Spanned, token::Comma, Expr, Ident, Token};
+use syn::{Expr, Ident, Token, parse::ParseBuffer, spanned::Spanned, token::Comma};
 
-use crate::load_locales::plurals::PluralForm;
+use leptos_i18n_codegen::load_locales::plurals::PluralForm;
 
 pub struct ParsedInput {
     pub context: Expr,
@@ -41,7 +41,10 @@ fn parse_plural_form(input: &ParseBuffer) -> syn::Result<(Option<PluralForm>, Ex
     } else if ident == "other" {
         PluralForm::Other
     } else {
-        return emit_err(ident, "Unknown form. Allowed forms are \"zero\", \"one\", \"two\", \"few\", \"many\", \"other\" and \"_\" fallback.");
+        return emit_err(
+            ident,
+            "Unknown form. Allowed forms are \"zero\", \"one\", \"two\", \"few\", \"many\", \"other\" and \"_\" fallback.",
+        );
     };
     Ok((Some(form), block, ident.span()))
 }
@@ -77,7 +80,7 @@ impl syn::parse::Parse for ParsedInput {
                     Some(PluralForm::Many) => "many",
                     Some(PluralForm::Other) => "other",
                 };
-                let msg = format!("Duplicate form {}.", form);
+                let msg = format!("Duplicate form {form}.");
                 return Err(syn::Error::new(span, msg));
             }
         }
