@@ -1,25 +1,15 @@
 #![deny(warnings)]
+#![cfg(test)]
 
-leptos_i18n::load_locales!();
+include!(concat!(env!("OUT_DIR"), "/i18n/mod.rs"));
 
-#[cfg(test)]
+mod components;
 mod defaulted;
-#[cfg(test)]
 mod foreign;
-#[cfg(test)]
 mod formatting;
-#[cfg(test)]
 mod plurals;
-#[cfg(test)]
-mod ranges;
-#[cfg(test)]
 mod scoped;
-#[cfg(test)]
 mod subkeys;
-#[cfg(test)]
-mod tests;
-
-#[cfg(test)]
 mod t_format;
-#[cfg(test)]
 mod t_plural;
+mod tests;
