@@ -1,7 +1,14 @@
 use crate::i18n::*;
-use leptos_i18n::formatting::*;
-use leptos_i18n::reexports::fixed_decimal::FixedDecimal;
-use leptos_i18n::reexports::icu::calendar::{Date, DateTime, Time};
+use leptos_i18n::{
+    formatting::*,
+    reexports::{
+        fixed_decimal::Decimal,
+        icu::{
+            calendar::Date,
+            datetime::input::{DateTime, Time},
+        },
+    },
+};
 use tests_common::*;
 
 #[test]
@@ -16,41 +23,45 @@ fn list_formatting() {
 
 #[test]
 fn date_formatting() {
-    let date = move || Date::try_new_iso_date(1970, 1, 2).unwrap().to_any();
+    let date = move || Date::try_new_iso(1970, 1, 2).unwrap().to_any();
 
     let en = td_format!(Locale::en, date, formatter: date);
     assert_eq_rendered!(en, "Jan 2, 1970");
-    let fr = td_format!(Locale::fr, date, formatter: date(date_length: full));
-    assert_eq_rendered!(fr, "vendredi 2 janvier 1970");
+    let fr = td_format!(Locale::fr, date, formatter: date(length: long));
+    assert_eq_rendered!(fr, "2 janvier 1970");
 }
 
 #[test]
 fn time_formatting() {
     let time = move || Time::try_new(14, 34, 28, 0).unwrap();
 
-    let en = td_format!(Locale::en, time, formatter: time);
+    let en = td_format!(Locale::en, time, formatter: time(time_precision: minute));
     assert_eq_rendered!(en, "2:34\u{202f}PM");
-    let fr = td_format!(Locale::fr, time, formatter: time);
-    assert_eq_rendered!(fr, "14:34");
+    let fr = td_format!(Locale::fr, time, formatter: time(time_precision: subsecond_s3));
+    assert_eq_rendered!(fr, "14:34:28,000");
 }
 
 #[test]
 fn datetime_formatting() {
     let date = move || {
-        let date = Date::try_new_iso_date(1970, 1, 2).unwrap().to_any();
+        let date = Date::try_new_iso(1970, 1, 2).unwrap().to_any();
         let time = Time::try_new(14, 34, 28, 0).unwrap();
-        DateTime::new(date, time)
+        DateTime { date, time }
     };
 
-    let en = td_format!(Locale::en, date, formatter: datetime);
+    let en = td_format!(Locale::en, date, formatter: datetime(time_precision: minute));
     assert_eq_rendered!(en, "Jan 2, 1970, 2:34\u{202f}PM");
-    let fr = td_format!(Locale::fr, date, formatter: datetime);
+    let fr = td_format!(Locale::fr, date, formatter: datetime(time_precision: minute));
     assert_eq_rendered!(fr, "2 janv. 1970, 14:34");
 }
 
 #[test]
 fn number_formatting() {
-    let num = move || FixedDecimal::from(200050).multiplied_pow10(-2);
+    let num = move || {
+        let mut dec = Decimal::from(200050);
+        dec.multiply_pow10(-2);
+        dec
+    };
 
     let en = td_format!(Locale::en, num, formatter: number);
     assert_eq_rendered!(en, "2,000.50");
